requirements.txt
google-genai
python-dotenv
tqdm
pyaml
urllib3<2
requests
python-docx
PyPDF2